epipyweb: import-syslog.py, uWSGI app and Javascript front end
Added import-syslog.py for importing all the DNS queries from
a syslog into the Epipyweb database.

When the DNS queries are stored in the database, they are grouped
by host if they occur in a short time range, with an extended range
for grouping used when the query value matches a query already in
the group.  When the queries are displayed, only the first query
value is shown, and additional connections are summarized by a count.

Added search-query.py for searching for DNS queries with a
particular value in the database.

Added a uWSGI app which retrieves DNS queries from the SQLite
database.

Added Javascript to build a list of DNS requests for display,
with "Next Page" and "Previous Page" links for paging through
the connection log.  Clicking the "X more connections" list
expands the group into a list including each DNS query made.