search: Added search bar
Added a text input for searching at the top of the HTML.  It will
find DNS queries which begin with the specified text and display
only groups containing those queries.

If part of a DNS query matches the search text, highlight
that part of the query value when listing DNS queries.

We check for the existence of data for the next and previous page
while collecting data for a page of DNS query groups, and pass
flags indicating the presence of such pages with the query groups.
The Javascript will only add the links to additional pages if the
flags are set.

Moved uWGI source to serve/
Moved HTML and Javascript to ui/
Moved syslog importer to record/

Added test.sh as a test runner

Added test/querygroup.py for creating a database from a simulated
syslog, and retrieving a query group via HTTP.

Configure rsyslogd to invoke episyslog.py when dnsmasq logs, and
record DNS queries and DCHP assignments to the database, which is
now located in /var/lib/epipyweb/dns.db