ui: added page tabs, including a new status page
Two page tabs have been added to the title bar:  Status and Connections.
Moved the connection display from the default (Status) tab to
the Connections tab.

The status page retrieves the device status from '/q/status'
and fill out a table with relevant IP addresses and SD card
space used on the status page.

If another DHCP device is present on the network, report that
device on the status page.  Provide a link to the device, and
suggest disabling DHCP on the other device.